Changed relation comprobation on Forum Added query for lastRequest of porfile on campaign Corrected possible let down from review ia commenter Added capaing request change message on kick
