+++ src/main/java/com/rolesync/rolesync/repository/CampaignRequestRepository.java
@@ -1,6 +1,7 @@
 package com.rolesync.rolesync.repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -12,17 +13,22 @@ import com.rolesync.rolesync.model.CampaignRequestStatus;
 import com.rolesync.rolesync.model.Profile;
 import com.rolesync.rolesync.repository.custominterfaces.CampaignRequestRepositoryCustom;
 
-public interface CampaignRequestRepository extends JpaRepository<CampaignRequest, Long>, QuerydslPredicateExecutor<CampaignRequest>, CampaignRequestRepositoryCustom{
-  
+public interface CampaignRequestRepository extends JpaRepository<CampaignRequest, Long>,
+        QuerydslPredicateExecutor<CampaignRequest>, CampaignRequestRepositoryCustom {
+
     List<CampaignRequest> findAllByCampaignAndStatus(Campaign campaign, CampaignRequestStatus status);
 
     List<CampaignRequest> findAllByCampaignAndProfile(Campaign campaign, Profile profile);
 
+    Optional<CampaignRequest> findTopByCampaignAndProfileOrderByIdDesc(
+            Campaign campaign,
+            Profile profile);
+
     @Query(value = """
-        SELECT COUNT(r)
-        FROM campaign_request r
-        WHERE r.campaign_id = :campaignId
-        AND r.status = 0
-        """, nativeQuery = true)
+            SELECT COUNT(r)
+            FROM campaign_request r
+            WHERE r.campaign_id = :campaignId
+            AND r.status = 0
+            """, nativeQuery = true)
     Long countPendingRequestsByCampaignId(Long campaignId);
 }
